Новый пользователь создается для каждого теста. Поля классов page objects приватные
