--- src/main/java/api/constants/EndPoints.java
@@ -1,9 +1,0 @@
-package api.constants;
-
-public final class EndPoints {
-    public static final String BASE_URL = "https://stellarburgers.nomoreparties.site";
-
-    public static final String CREATE_USER_PATH = "/api/auth/register";
-    public static final String LOGIN_USER_PATH = "/api/auth/login";
-    public static final String REMOVE_USER_PATH = "/api/auth/user";
-}
