--- src/main/java/ui/pageobjects/RecoverPasswordPage.java
@@ -1,19 +1,0 @@
-package ui.pageobjects;
-
-import io.qameta.allure.Step;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-
-public class RecoverPasswordPage extends Page{
-
-    public static final By LOGIN_LINK = By.linkText("Войти");
-
-    public RecoverPasswordPage(WebDriver driver) {
-        super(driver);
-    }
-
-    @Step("Нажать на ссылку \"Войти\"")
-    public void clickLoginLink(){
-        driver.findElement(LOGIN_LINK).click();
-    }
-}
